refactor: merge material supplier transformers into material_transformer.go
Consolidates ToMaterialSupplier* functions from the separate
material_supplier_transformer.go into material_transformer.go,
matching the same pattern used for the entity structs.
